Refactor to project structure. Added Terrain and Area Type steps
