+++ app/src/main/java/com/example/autosar/composables/subjectWizard/models/WizardStep.kt
@@ -1,0 +1,3 @@
+package com.example.autosar.composables.subjectWizard.models
+
+enum class WizardStep { Category, Activity, Terrain, AreaType }
